Player Can Now Move
The capsule that is the player can now move around the environment without clipping. It also can't fall off the map, nor climb mountains. There is one hidden spot that still can act as a shield for the player, but it can either be easily reversed or turned into an enemy spawn point.

The interface used for movement is WASD to move and mouse to look around.